+++ lib/decidim/transparent_trash/engine.rb
@@ -22,6 +22,7 @@ module Decidim
       initializer "decidim.transparent_trash.extends" do
         Decidim::Initiatives::Admin::UnpublishInitiative.include Decidim::TransparentTrash::Extends::UnpublishInitiative
         Decidim::Initiatives::Permissions.include Decidim::TransparentTrash::Extends::InitiativesPermissions
+        Decidim::Initiatives::Admin::Permissions.include Decidim::TransparentTrash::Extends::InitiativesAdminPermissions
       end
     end
   end
